Added Swal alert to delete item from cart

+++ src/app/components/cart/cart.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { ProductService } from "../../services/product.service";
 import { ActivatedRoute, Router } from "@angular/router";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "app-cart",
@@ -33,12 +34,23 @@ export class CartComponent implements OnInit {
   }
 
   delete(cartId) {
-    console.log(cartId);
-    this.productService.deleteCartItem(cartId).subscribe(
-      (resp) => {
-        this.getCartDetails();
-      },
-      (err) => {}
-    );
+    Swal.fire({
+      title: "Do you want to delete item from cart?",
+      showDenyButton: true,
+      confirmButtonText: "Delete Item",
+      denyButtonText: `Don't Delete`
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productService.deleteCartItem(cartId).subscribe(
+          (resp) => {
+            this.getCartDetails();
+          },
+          (err) => {
+            console.log(err);
+          }
+        );
+        Swal.fire("Done!", "", "success");
+      }
+    });
   }
 }
